+++ lib/types/betting_coefficient.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module Types
+  BettingCoefficient = Types::String.constructor { _1 || '1' }
+end