+++ spec/rails_helper.rb
@@ -14,6 +14,7 @@ require 'support/factory_bot'
 require 'support/tasks'
 require 'support/active_admin'
 require 'support/time'
+require 'support/sidekiq'
 
 # Requires supporting ruby files with custom matchers and macros, etc, in
 # spec/support/ and its subdirectories. Files matching `spec/**/*_spec.rb` are