+++ spec/support/sidekiq.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require 'sidekiq/testing'
+
+Sidekiq::Testing.inline!