+++ gems/oddsportal_scraper/CHANGELOG.md
@@ -11,3 +11,11 @@
 ## [0.2.1] - 2022-10-30
 
 - Matches scraper can accept a new format for the season (`2020/2021`)
+
+## [0.2.2] - 2022-11-19
+
+- Return empty response for invalid urls.
+
+## [0.2.3] - 2022-11-22
+
+- Return empty response for season without bookmakers odds.